Read in a text file generated by Eprime

Read in a text file generated by Eprime

read_eprime(filename, remove_clock = TRUE)

Arguments

filename
Either the full or relative path to an Eprime .txt file

remove_clock
Whether to exclude the Clock.Information XML entries. Enabled by default.

Value

Each line of the file is stored and returned in a character vector.

Details

The encoding on an Eprime txt file should be UCS-2 Little Endian, but sometimes this is not the case. We delegate the fussy encoding details to the stringi::str_read_lines function.

If the file is not an Eprime txt--that is, if it is missing the lines *** Header Start *** and *** Header End ***--a warning is raised and the lines of text are replaced by a dummy header.